# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  utils.Result:
    properties:
      data: {}
      message:
        type: string
      statusCode:
        type: integer
      success:
        type: boolean
      systemMessage:
        type: string
    type: object
  website.GetSingleArticleBySlugQuery:
    properties:
      domain:
        type: string
      siteId:
        type: integer
      slug:
        type: string
    required:
    - domain
    - slug
    type: object
  website.GetSingleProductBySlugQuery:
    properties:
      domain:
        type: string
      siteId:
        type: integer
      slug:
        type: string
    required:
    - domain
    - slug
    type: object
host: localhost:8585
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: This is a sample server.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://example.com/terms/
  title: My API
  version: "1.0"
paths:
...
  /website/article:
    get:
      consumes:
      - application/json
      description: Retrieves a specific article identified by its slug
      parameters:
      - description: Article slug
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/website.GetSingleArticleBySlugQuery'
      produces:
      - application/json
      responses:
        "200":
          description: Article details
          schema:
            $ref: '#/definitions/utils.Result'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/utils.Result'
        "404":
          description: Article not found
          schema:
            $ref: '#/definitions/utils.Result'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Result'
      summary: Get article by slug
      tags:
      - website
  /website/article/category:
    get:
      consumes:
      - application/json
      description: Retrieves articles belonging to a specific category identified
        by slug
      parameters:
      - in: query
        name: domain
        required: true
        type: string
      - in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: pageSize
        required: true
        type: integer
      - in: query
        name: search
        type: string
      - enum:
        - title
        - description
        in: query
        name: searchBy
        type: string
      - in: query
        name: siteId
        type: integer
      - in: query
        name: slug
        required: true
        type: string
      - enum:
        - asc
        - desc
        in: query
        name: sort
        type: string
      - enum:
        - title
        - description
        in: query
        name: sortBy
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of articles in the category
          schema:
            $ref: '#/definitions/utils.Result'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/utils.Result'
        "404":
          description: Category not found
          schema:
            $ref: '#/definitions/utils.Result'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Result'
      summary: Get articles by category slug
      tags:
      - website
  /website/article/filters-sort:
    get:
      consumes:
      - application/json
      description: Retrieves available filters and sorting options for articles
      parameters:
      - in: query
        name: domain
        required: true
        type: string
      - in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: pageSize
        required: true
        type: integer
      - in: query
        name: search
        type: string
      - enum:
        - title
        - description
        in: query
        name: searchBy
        type: string
      - enum:
        - title_a_z
        - title_z_a
        - recently_added
        - recently_updated
        - most_visited
        - least_visited
        - most_rated
        - least_rated
        - most_reviewed
        - least_reviewed
        in: query
        name: selectedSort
        type: string
        x-enum-varnames:
        - ArticleTitleAZSort
        - ArticleTitleZASort
        - ArticleRecentlyAddedSort
        - ArticleRecentlyUpdatedSort
        - ArticleMostVisitedSort
        - ArticleLeastVisitedSort
        - ArticleMostRatedSort
        - ArticleLeastRatedSort
        - ArticleMostReviewedSort
        - ArticleLeastReviewedSort
      - in: query
        name: siteId
        type: integer
      - enum:
        - asc
        - desc
        in: query
        name: sort
        type: string
      - enum:
        - title
        - description
        in: query
        name: sortBy
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Article filters and sorting options
          schema:
            $ref: '#/definitions/utils.Result'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/utils.Result'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Result'
      summary: Get article filters and sorting options
      tags:
      - website
  /website/header-footer:
    get:
      consumes:
      - application/json
      description: Retrieves header and footer content for a specific domain
      parameters:
      - in: query
        name: domain
        required: true
        type: string
      - in: query
        name: siteId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Header and footer content
          schema:
            $ref: '#/definitions/utils.Result'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/utils.Result'
        "404":
          description: Domain not found
          schema:
            $ref: '#/definitions/utils.Result'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Result'
      summary: Get header and footer by domain
      tags:
      - website
  /website/page:
    get:
      consumes:
      - application/json
      description: Retrieves page content for a specific domain
      parameters:
      - in: query
        name: domain
        required: true
        type: string
      - in: query
        name: path
        type: string
      - in: query
        name: siteId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Page content
          schema:
            $ref: '#/definitions/utils.Result'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/utils.Result'
        "404":
          description: Page or domain not found
          schema:
            $ref: '#/definitions/utils.Result'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Result'
      summary: Get page by domain
      tags:
      - website
  /website/product:
    get:
      consumes:
      - application/json
      description: Retrieves a specific product identified by its slug
      parameters:
      - description: Product slug
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/website.GetSingleProductBySlugQuery'
      produces:
      - application/json
      responses:
        "200":
          description: Product details
          schema:
            $ref: '#/definitions/utils.Result'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/utils.Result'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/utils.Result'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Result'
      summary: Get product by slug
      tags:
      - website
  /website/product/category:
    get:
      consumes:
      - application/json
      description: Retrieves products belonging to a specific category identified
        by slug
      parameters:
      - in: query
        name: domain
        required: true
        type: string
      - in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: pageSize
        required: true
        type: integer
      - in: query
        name: search
        type: string
      - enum:
        - title
        - description
        in: query
        name: searchBy
        type: string
      - in: query
        name: siteId
        type: integer
      - in: query
        name: slug
        required: true
        type: string
      - enum:
        - asc
        - desc
        in: query
        name: sort
        type: string
      - enum:
        - title
        - description
        in: query
        name: sortBy
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of products in the category
          schema:
            $ref: '#/definitions/utils.Result'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/utils.Result'
        "404":
          description: Category not found
          schema:
            $ref: '#/definitions/utils.Result'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Result'
      summary: Get products by category slug
      tags:
      - website
  /website/product/filters-sort:
    get:
      consumes:
      - application/json
      description: Retrieves available filters and sorting options for products
      parameters:
      - in: query
        name: domain
        required: true
        type: string
      - in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: pageSize
        required: true
        type: integer
      - in: query
        name: search
        type: string
      - enum:
        - title
        - description
        in: query
        name: searchBy
        type: string
      - enum:
        - price_low_to_high
        - price_high_to_low
        - coupon_high_to_low
        - name_a_z
        - name_z_a
        - recently_added
        - recently_updated
        - most_selling
        - most_visited
        - most_rated
        - most_reviewed
        - least_visited
        - least_rated
        - least_reviewed
        in: query
        name: selectedSort
        type: string
        x-enum-varnames:
        - ProductPriceLowToHighSort
        - ProductPriceHighToLowSort
        - ProductCouponHighToLowSort
        - ProductNameAZSort
        - ProductNameZASort
        - ProductRecentlyAddedSort
        - ProductRecentlyUpdatedSort
        - ProductMostSellingSort
        - ProductMostVisitedSort
        - ProductMostRatedSort
        - ProductMostReviewedSort
        - ProductLeastVisitedSort
        - ProductLeastRatedSort
        - ProductLeastReviewedSort
      - in: query
        name: siteId
        type: integer
      - enum:
        - asc
        - desc
        in: query
        name: sort
        type: string
      - enum:
        - title
        - description
        in: query
        name: sortBy
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product filters and sorting options
          schema:
            $ref: '#/definitions/utils.Result'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/utils.Result'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Result'
      summary: Get product filters and sorting options
      tags:
      - website
  /website/product/search:
    get:
      consumes:
      - application/json
      description: Searches for products based on provided criteria
      parameters:
      - in: query
        name: domain
        required: true
        type: string
      - in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: pageSize
        required: true
        type: integer
      - in: query
        name: search
        type: string
      - enum:
        - title
        - description
        in: query
        name: searchBy
        type: string
      - in: query
        name: siteId
        type: integer
      - enum:
        - asc
        - desc
        in: query
        name: sort
        type: string
      - enum:
        - title
        - description
        in: query
        name: sortBy
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of products matching search criteria
          schema:
            $ref: '#/definitions/utils.Result'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/utils.Result'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Result'
      summary: Search products
      tags:
      - website
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"